backend/requirements.txt
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
